\name{drop_auxiliary}
\alias{drop_auxiliary}
\title{drop auxiliary values}
\usage{
drop_auxiliary(tt, destructive = FALSE)
}
\arguments{
  \item{tt}{time.table to drop auxiliary columns from}

  \item{destructive}{wether to mutate existing time table,
  defaults to false}
}
\description{
drop auxiliary values
}

