\name{integer_index}
\alias{integer_index}
\title{Create an integer index}
\usage{
integer_index(tt, new.index.name = "index")
}
\arguments{
  \item{tt}{time.table to replace index of}

  \item{new.index.name}{name of new index column, defaults
  to "index"}
}
\description{
Create a single column integer index to replace the current
one.
}

