\name{integer_time}
\alias{integer_time}
\title{Create an integer time}
\usage{
integer_time(tt, new.time.name = "time")
}
\arguments{
  \item{tt}{time.table to replace time of}

  \item{new.time.name}{name of new time column, defaults to
  "time"}
}
\description{
Create a (discrete) integer based time column
}

