\name{nobs}
\alias{nobs}
\title{number of observations}
\usage{
nobs(tt, units = FALSE, theoretical = FALSE)
}
\arguments{
  \item{tt}{time.table the number of observations of which
  to query}

  \item{units}{whether to ask for te number of (cross
  sectional) units rather than the number of unit/time
  pairs, defaults to the latter}

  \item{theoretical}{whether to compute the number of
  observations based on frequency information or on actual
  data stored in the time.table, defaults to the latter}
}
\description{
number of observations
}

